+++ ChangeLog.md
@@ -16,6 +16,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Filesystem::chmod()
 - Filesystem::utime()
 - Filesystem::write()
+- Filesystem::link()
+- Filesystem::symlink()
+- Filesystem::rename()
 
 ## [0.0.3] - 2024-10-25